refactor: move notes config to shared/config.ts, parameterize sanitizeBodyContent
- Move getNotesFolders(), getDefaultNotesAccount(), getNotesDbPath() from
  notes/tools.ts to shared/config.ts alongside other module configs.
  Re-exported from notes/tools.ts for backward compatibility.
- Add source parameter to sanitizeBodyContent() (default 'EMAIL').
  Notes passes 'NOTE' for accurate content labeling.
- Add test for custom sanitizeBodyContent source label.
